improve(PriceClient): Rename across -> acrossApi
Feedback from Matt.
