+++ src/priceClient/adapters/index.ts
@@ -1,2 +1,2 @@
-export * as across from "./across";
+export * as acrossApi from "./acrossApi";
 export * as coingecko from "./coingecko";
